requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-admin-action-buttons==0.1.9
idna==3.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
